expense_tracker/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-storages==1.14.3
django-sendgrid-v5==1.2.3
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
Pillow==10.3.0
boto3==1.34.69
python-decouple==3.8
factory-boy==3.3.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
flower==2.0.1
sentry-sdk==1.44.1
django-extensions==3.2.3
